+++ src/App/Config/Routes.php
@@ -44,7 +44,7 @@ function registerRoutes(App $app)
     $app->get('/logout', [AuthController::class, 'logout'], false)
         ->add([AuthRequiredMiddleware::class]);
 
-    $app->get('/friends', [FriendsController::class, 'getFriends'], true) // not implemented yet
+    $app->get('/friends', [FriendsController::class, 'getFriends'], true)
         ->add([AuthRequiredMiddleware::class]);
 
     $app->post('/friends', [FriendsController::class, 'sendRequest'], true) // good
@@ -84,6 +84,6 @@ function registerRoutes(App $app)
     $app->post('/api/posts/', [PostsController::class, 'addComment'], true)
         ->add([AuthRequiredMiddleware::class]);
 
-    $app->post('/api/reports', [ReportController::class, 'makeReport'], true)
+    $app->post('/api/reports', [ReportController::class, 'sendReport'], true)
         ->add([AuthRequiredMiddleware::class]);
 }
